Reward

+++ WhoWantToBeAMillionaire/Models/PlayerModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string UserId { get; set; }
         public string Name { get; set; }
-        public int Score { get; set; } = 0;
+        public decimal Score { get; set; } = 0;
     }
 
 }
